direct illumination

+++ rengine-src/Rengine/Renderer/Material.hpp
@@ -7,6 +7,6 @@ namespace Rengin
 struct Material
 {
     glm::vec3 Ks,Kd,Ka,Le;
-    double Ns,Ni;
+    float Ns,Ni;
 };
 } // namespace Rengin
